Python: moved AI/Python.cpp and .h to python/AI/
